+++ generators/aws-swarm-leader/templates/aws-swarm-leader/config.ts
@@ -60,44 +60,44 @@ export const getConfig = async () => {
     keypair = outputs ? outputs[0]["name"] as string : undefined;
   }
 
-  const pathToSshKeysFolder = stackConfig.get("pathToSshKeysFolder") || "../../../ssh-keys";
+  /** Gets security group ids */
+  const useBastion = stackConfig.getBoolean("useBastion");
 
-  const publicKeyNames = stackConfig.requireObject("publicKeyNames") as string[];
-
-  /** Gets security group id */
   let securityGroupIds = stackConfig.getObject<string[]>("securityGroupIds");
 
   if (!securityGroupIds) {
-    const outputs = await getOutputs(
-      "securityGroupStack",
-      "managersSecurityGroupId"
+    const securityGroupNames = useBastion
+      ? "swarm-managers,web,ssh-bastion"
+      : "swarm-managers,web";
+
+    const outputs = await getOutputs<{ "arn": string; "id": string }>(
+      "securityGroupsStack",
+      securityGroupNames
     );
 
-    securityGroupIds = outputs ? [outputs[0] as string] : [];
+    if (!outputs) {
+      throw new Error("Required security group could not be found");
+    }
+
+    try {
+      const managers = outputs[0] as { "arn": string; "id": string };
+      const web = outputs[1] as { "arn": string; "id": string };
+
+      securityGroupIds = [managers["id"], web["id"]];
+
+      if (useBastion) {
+        const bastion = outputs[2] as { "arn": string; "id": string };
+        securityGroupIds.push(bastion["id"]);
+      }
+    } catch (e) {
+      throw new Error("Required security groups could not be found");
+    }
   }
 
   /** Get subnet id **/
   const subnetId = stackConfig.require("subnetId");
 
-  const useNfs = stackConfig.getBoolean("useNfs");
-
-  /** Get user data **/
-  const userData = generateUserData(
-    stackConfig.get("userDataTemplate") || "./cloud-config.al2023.njx",
-    {
-      dockerNetworks: stackConfig.getObject<string[]>("dockerNetworks"),
-      packages: stackConfig.getObject<string[]>("packages"),
-      publicKeyNames: getPublicKeys(publicKeyNames, pathToSshKeysFolder),
-      volumes: [
-        {
-          device: stackConfig.get("volumeDevice") || "/dev/xvdf",
-          filesystem: stackConfig.get("volumeFilesystem") || "ext4",
-          label: stackConfig.get("volumeLabel") || "data",
-          path: "/mnt/data"
-        }
-      ],
-    }
-  );
+  const useNFS = stackConfig.getBoolean("useNFS");
 
   /** Get volume id **/
   let volumeId = stackConfig.get("volumeId");
@@ -110,6 +110,42 @@ export const getConfig = async () => {
 
     volumeId = outputs ? outputs[0] as string : undefined;
   }
+
+  /** Get VPC id */
+  let vpcId = stackConfig.get("vpcId");
+
+  if (!vpcId) {
+    const outputs = await getOutputs(
+      "vpcStack",
+      "vpcId"
+    );
+
+    if (outputs) {
+      vpcId = outputs[0] as string;
+    }
+  }
+
+  /** User data **/
+  const pathToSshKeysFolder = stackConfig.get("pathToSshKeysFolder") || "../../../ssh-keys";
+
+  const publicKeyNames = stackConfig.requireObject("publicKeyNames") as string[];
+
+  const userData = generateUserData(
+    stackConfig.get("userDataTemplate") || "./cloud-config.al2023.njx",
+    {
+      dockerNetworks: stackConfig.getObject<string[]>("dockerNetworks"),
+      packages: stackConfig.getObject<string[]>("packages"),
+      publicKeyNames: getPublicKeys(publicKeyNames, pathToSshKeysFolder),
+      volumes: useNFS ? undefined : [
+        {
+          device: stackConfig.get("volumeDevice") || "/dev/xvdf",
+          filesystem: stackConfig.get("volumeFilesystem") || "ext4",
+          label: stackConfig.get("volumeLabel") || "data",
+          path: "/mnt/data"
+        }
+      ],
+    }
+  );
 
   return {
     ami: stackConfig.require("ami"),
@@ -131,9 +167,11 @@ export const getConfig = async () => {
     securityGroupIds,
     subnetId,
     tags: stackConfig.getObject<{ [key: string]: string }>("tags"),
-    useNfs,
+    useBastion,
+    useNFS,
     userData,
     volumeId,
+    vpcId,
   };
 };